PSJ-CS-Web/psj-cs-api/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
mysqlclient
SQLAlchemy
flask_cors
mysql-connector-python
